pyproject.toml
[project]
name = "gently-annotator"
version = "0.1.0"
description = "Standalone web app for annotating C. elegans developmental stages on diSPIM volumes."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "tifffile>=2024.1",
    "numpy>=1.24",
    "scipy>=1.11",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "httpx>=0.26",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["annotator"]
